Phase 0: carry RR intervals as unrounded Double milliseconds
HeartRateCodec rounded RR from 1/1024 s to integer ms, then the pipeline
re-widened to Double — ~0.5 ms quantisation riding on RMSSD. Carry RR as
fractional ms (Double) from the codec through HeartRate, the pipeline sample,
and the CSV row. The dashboard readout keeps integer ms; the analysed and
recorded paths keep full precision.

+++ Sources/DeviceCore/HeartRate.swift
@@ -6,9 +6,11 @@ import Foundation
 /// skin-contact support.
 public struct HeartRate: Sendable, Equatable {
     public let bpm: Int
-    public let rrIntervalsMs: [Int]
+    /// RR intervals in milliseconds, unrounded — the raw 1/1024 s units carried as
+    /// fractional ms so ~0.5 ms quantisation doesn't ride on downstream HRV.
+    public let rrIntervalsMs: [Double]
     public let contact: Bool?
-    public init(bpm: Int, rrIntervalsMs: [Int], contact: Bool?) {
+    public init(bpm: Int, rrIntervalsMs: [Double], contact: Bool?) {
         self.bpm = bpm
         self.rrIntervalsMs = rrIntervalsMs
         self.contact = contact
